# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Find Your Car",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/fyc/zoneState": {
            "put": {
                "description": "Change the state of a Zone (e.g., enabled/disabled) or retrieve a Zone by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Change Zone state or retrieve Zones by ID",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Zone State",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Zone ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Number of rows affected by the state change",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/fyc/zones": {
            "get": {
                "description": "Get a list of all zones, or a zone by ID if 'id' parameter is provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Get all zones",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Include extra information if 'yes'",
                        "name": "extra",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Zone ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of zones or a single zone",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.Zone"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new zone to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Add a new zone",
                "parameters": [
                    {
                        "description": "Zone data",
                        "name": "zone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Zone"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pkg.Zone"
                        }
                    }
                }
            }
        },
        "/fyc/zones/{id}": {
            "put": {
                "description": "Update an existing zone by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Update a zone by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Zone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated zone data",
                        "name": "zone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Zone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Zone"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a zone by ID",
                "tags": [
                    "Zones"
                ],
                "summary": "Delete a zone",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Zone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Zone not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/fyc/zonesDeleted": {
            "get": {
                "description": "Get a list of Deleted zones or a specific zone by ID with optional extra data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Get Deleted zones or a specific zone by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Deleted zones or a single zone",
                        "schema": {
                            "$ref": "#/definitions/pkg.Zone"
                        }
                    }
                }
            }
        },
        "/fyc/zonesEnabled": {
            "get": {
                "description": "Get a list of enabled zones or a specific zone by ID with optional extra data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Zones"
                ],
                "summary": "Get enabled zones or a specific zone by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of enabled zones or a single zone",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.Zone"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pkg.Zone": {
            "type": "object",
            "required": [
                "extra",
                "free_capacity",
                "last_update",
                "max_capacity",
                "name",
                "zone_id"
            ],
            "properties": {
                "extra": {
                    "type": "object"
                },
                "free_capacity": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean",
                    "default": false
                },
                "is_enabled": {
                    "type": "boolean",
                    "default": false
                },
                "last_update": {
                    "type": "string"
                },
                "max_capacity": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "zone_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}